Move sparkctl to cmd directory
