+++ cmd/sparkctl/app/testdata/hadoop-conf/binary.dat
