+++ cmd/sparkctl/main.go
@@ -1,5 +1,5 @@
 /*
-Copyright 2018 Google LLC
+Copyright 2017 Google LLC
 
 Licensed under the Apache License, Version 2.0 (the "License");
 you may not use this file except in compliance with the License.
@@ -14,26 +14,14 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package cmd
+package main
 
 import (
-	"time"
+	_ "k8s.io/client-go/plugin/pkg/client/auth"
 
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/apimachinery/pkg/util/duration"
+	"github.com/kubeflow/spark-operator/cmd/sparkctl/app"
 )
 
-func getSinceTime(timestamp metav1.Time) string {
-	if timestamp.IsZero() {
-		return "N.A."
-	}
-
-	return duration.ShortHumanDuration(time.Since(timestamp.Time))
-}
-
-func formatNotAvailable(info string) string {
-	if info == "" {
-		return "N.A."
-	}
-	return info
+func main() {
+	app.Execute()
 }
